add - Added a huge amount of names!
---

We've added a huge amount of first names and surnames from a data dump found publicly at https://github.com/philipperemy/name-dataset. Thanks to @philipperemy for this repository!

As for the sponsorship, current restrictions forbid us from sponsoring that repository. Appropriate credits are placed in the README.md file instead.

Now, the names list has grown from 70,000+ names and 100,000+ surnames to 700,000+ names (common and rare) and 900,000+ surnames (again common and rare). This is to make sure that this names list is going to be the most complete names list (that are written in English) accessible from just plain text!

---

Type: add
Breaking: False
Doc Required: False
Part: 1/1